+++ include/slab.h
@@ -17,15 +17,15 @@ enum transfer_vals {
 typedef struct slab_header_s {
 	unsigned magic;
 	unsigned id;
-	unsigned short num_pages;
+	unsigned num_pages;
 	unsigned short flags;
-	unsigned short obj_used;/* Ref count */
-	unsigned short obj_num;
+	unsigned obj_used;/* Ref count */
+	unsigned obj_num;
 	/* For usage in the slab lists */
 	struct slab_header_s *next, *prev;
 	addr_t parent; /* Pointer to the slab cache that this slab is part of */
-	unsigned short *stack;
-	unsigned short stack_arr[MAX_OBJ_ID];
+	unsigned *stack;
+	unsigned stack_arr[MAX_OBJ_ID];
 	vnode_t *vnode;
 } slab_t;
